Refactor of flask application
Refactor in folders, libraries and the way we program the code
