+++ resources/Flask/app/static/css/loadingStyles.css
