+++ resources/Flask/app/static/css/styles.css
