+++ resources/Flask/app/static/js/indexManual.js
