+++ resources/Flask/app/static/js/keyboardLayout.js
