+++ resources/Flask/app/static/js/scripts.js
