+++ resources/Flask/app/static/js/sidebar.js
