+++ resources/Flask/app/static/js/switchButtons.js
