+++ resources/Flask/app/static/js/validarForm.js
