+++ resources/Flask/app/templates/automatico.html
