+++ resources/Flask/app/templates/includes/humTemp.html
