+++ resources/Flask/app/templates/includes/navbar.html
@@ -51,7 +51,7 @@
     </ul>
     <div class="rights">
       Administrador de terrario, todos los derechos reservados -
-      {{ now.day }}/{{now.month }}/{{ now.year }}
+      {{ now.day }}/{{ now.strftime("%B") }}/{{ now.year }}
     </div>
   </div>
 </nav>
