+++ resources/Flask/app/templates/includes/scripts.html
