+++ resources/Flask/app/templates/manual.html
